Don't print warning about dot not being installed from the toplevel
By checking if dot is installed from the toplevel of a loadable
module, code that use the module will get a warning about dot not
being installed, even if the user code has nothing to do with graph
generation (e.g., salmonella-feeds).

While at it, turn `dot-installed?' into a procedure.